DigiTriadLab/app/src/main/java/com/rfdetke/digitriadlaboratory/export/csv/BluetoothLeCsvFileWriter.java
package com.rfdetke.digitriadlaboratory.export.csv;

import android.content.Context;

import com.rfdetke.digitriadlaboratory.constants.SourceTypeEnum;
import com.rfdetke.digitriadlaboratory.database.AppDatabase;
import com.rfdetke.digitriadlaboratory.repositories.BluetoothLeRepository;

import java.util.ArrayList;
import java.util.List;

public class BluetoothLeCsvFileWriter extends CsvSampleFileWriter {

    public BluetoothLeCsvFileWriter(long[] runIds, AppDatabase database, Context context) throws NullPointerException {
        super(runIds, database, context);
    }

    @Override
    public List<CsvExportable> getExportableData() {
        return new ArrayList<>(new BluetoothLeRepository(database).getAllSamples(runIds));
    }

    @Override
    public String getKey() {
        return SourceTypeEnum.BLUETOOTH_LE.name();
    }
}
